requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py39hecd8cb5_1001
backcall=0.2.0=pyhd3eb1b0_0
ca-certificates=2021.7.5=hecd8cb5_1
certifi=2021.5.30=py39hecd8cb5_0
decorator=5.0.9=pyhd3eb1b0_0
ipython=7.26.0=py39h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py39hecd8cb5_1
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
matplotlib-inline=0.1.2=pyhd3eb1b0_2
ncurses=6.2=h0a44026_1
openssl=1.1.1l=h9ed2024_0
parso=0.8.2=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=21.2.4=py37hecd8cb5_0
prompt-toolkit=3.0.17=pyhca03da5_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pygments=2.10.0=pyhd3eb1b0_0
pyjwt=2.1.0=py39hecd8cb5_0
python=3.9.6=h88f2d9e_1
readline=8.1=h9ed2024_0
setuptools=52.0.0=py39hecd8cb5_0
sqlite=3.36.0=hce871da_0
tk=8.6.10=hb0a8c7a_0
traitlets=5.0.5=pyhd3eb1b0_0
tzdata=2021a=h5d7bf9c_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
